Delete manual editing of SubCategory change in maincategories.
This is because we are not making copies of subcategories when
attach them to maincategories. They are attached by reference.
This means that when we modify the values of a subcategory outside
of a Budget, this change gets repercuted in this.subcategories, but
also in the parent MainCategory of the modified subcategory