Add more precision for date in datePicker
 This is done to be able to sort the transactions
by date. If they are inserted on another
date than the current date, the hours,
minutes and seconds are not taken
into account in datePicker,
which made it that transactions entered on a date prior to DateTime.now()
were not sorted properly.
Now, we complete the year, month and day selected
with the insertion hour, minute and second so that
they are in sequential insert order.
#14