Fix payees not being updated after creation
After adding a payee using 'Create payee', although the new payee
was saved in the AppState, this change was not visible when selecting
a new payee again because the list of payees (and accounts) sent
to SelectValue are done in initState. Now, it is done in handleOnTapPayee,
which makes it so that the payees and accounts are fetched from
AppState every time we click on the Payee field.